+++ src/env-var/types.d.ts
@@ -1,4 +1,4 @@
-export type ContextType = "env-var" | "env-var-simple" | "env-var-addon" | "exposed-config";
+export type ContextType = "env-var" | "env-var-simple" | "env-var-addon" | "exposed-config" | "config-provider";
 
 export type ErrorType = "saving" | "loading";